+++ src/hypercube.cpp
@@ -10,13 +10,13 @@ void HyperCube::generateRandomPoint(std::vector<double>& random_point)
     std::uniform_real_distribution<double> distribution(-edge / 2, edge / 2);
 
 #pragma omp parallel for
-    for (int i = 0; i < dimension; ++i)
+    for (size_t i = 0; i < dimension; ++i)
         random_point[i] = distribution(eng);
 }
 
 void HyperCube::calculateVolume()
 {
-    for (int i = 0; i < dimension; ++i)
+    for (size_t i = 0; i < dimension; ++i)
     {
         volume *= edge;
     }
